refactor: CardScreen renamed to ParkingDetailsScreen
